+++ src/basic/components/index.js
@@ -1,20 +1,17 @@
 export * from './Header';
 export * from './GridContainer';
 
-// 좌측 아이템 선택 영역
-export * from './SelectItemSection/SelectorContainer';
 export * from './SelectItemSection/SelectItem';
-export * from './SelectItemSection/AddToCartBtn';
-export * from './SelectItemSection/CartItem';
-export * from './SelectItemSection/CartDisplay';
+export * from './SelectItemSection/SelectorContainer';
 export * from './SelectItemSection/ProductDropdown';
 export * from './SelectItemSection/ProductDropdownOptions';
+export * from './SelectItemSection/CartDisplay';
+export * from './SelectItemSection/CartItem';
+export * from './SelectItemSection/AddButton';
 export * from './SelectItemSection/StockInfo';
 
-// 우측 주문 요약 영역
 export * from './OrderSummaySection/OrderSummary';
 
-// 매뉴얼 영역
 export * from './ManualSection/Manual';
 export * from './ManualSection/ManualToggle';
 export * from './ManualSection/ManualOverlay';
