+++ src/basic/utils/index.js
@@ -1,1 +1,5 @@
+export * from './calculationUtils';
 export * from './cartUtils';
+export * from './domUtils';
+export * from './eventHandlers';
+export * from './eventTimers';
